requirements.txt
langgraph
openai
python-dotenv
langchain
langchain_openai
ipython
langchain-community
faiss-cpu
langgraph-cli
langgraph-cli[inmem]